Add a script to pull down chapter and add it as a draft
This is a script that when run will download the desired chapter from Syosetsu
and will reformat it into the markdown specific format that is used for this
site. Created specifically for the Honzuki story, but may work for others if
the ncode in the script is changed. After it's pulled down and reformatted,
it's added as a new draft that is ready to be translated, edited, and finally
published once complete.